2. _calloc

+++ 0x0C-more_malloc_free/2-calloc.c
@@ -17,7 +17,7 @@ void *_calloc(unsigned int nmemb, unsigned int size)
 	p = malloc(nmemb * size);
 	if (!p)
 		return (NULL);
-	for (i = 0; i < nmemb * size; i++)
+	for (i = 0; (unsigned int)i < nmemb * size; i++)
 		p[i] = 0;
 	return (p);
 }
